kernel: syscalls internal implementation inside an object
Syscall internal implementations were spread across headers, as static
inline, and C sources. The generated syscall code was declaring all
internal _impl_ as an extern function because of the ones implemented
in C sources. This was violating MISRA-C rule 8.8 in the cases
internal functions were implemented in the headers.

This patch put all internal implementations inside C sources.

MISRA-C rule 8.8
