+++ include/can.h
@@ -271,14 +271,6 @@ struct can_driver_api {
 __syscall int can_send(struct device *dev, struct can_msg *msg,
 		       s32_t timeout, can_tx_callback_t callback_isr);
 
-static inline int _impl_can_send(struct device *dev, struct can_msg *msg,
-				 s32_t timeout, can_tx_callback_t callback_isr)
-{
-	const struct can_driver_api *api = dev->driver_api;
-
-	return api->send(dev, msg, timeout, callback_isr);
-}
-
 /*
  * Derived can APIs -- all implemented in terms of can_send()
  */
@@ -327,48 +319,13 @@ static inline int can_write(struct device *dev, u8_t *data, u8_t length,
 __syscall int can_attach_msgq(struct device *dev, struct k_msgq *msg_q,
 			      const struct can_filter *filter);
 
-static inline int _impl_can_attach_msgq(struct device *dev,
-					struct k_msgq *msg_q,
-					const struct can_filter *filter)
-{
-	const struct can_driver_api *api = dev->driver_api;
-
-	return api->attach_msgq(dev, msg_q, filter);
-}
-
-
 __syscall int can_attach_isr(struct device *dev, can_rx_callback_t isr,
 			     const struct can_filter *filter);
-static inline int _impl_can_attach_isr(struct device *dev,
-				       can_rx_callback_t isr,
-				       const struct can_filter *filter)
-{
-	const struct can_driver_api *api = dev->driver_api;
-
-	return api->attach_isr(dev, isr, filter);
-}
-
 
 __syscall void can_detach(struct device *dev, int filter_id);
 
-static inline void _impl_can_detach(struct device *dev, int filter_id)
-{
-	const struct can_driver_api *api = dev->driver_api;
-
-	return api->detach(dev, filter_id);
-}
-
-
 __syscall int can_configure(struct device *dev, enum can_mode mode,
 			    u32_t bitrate);
-
-static inline int _impl_can_configure(struct device *dev, enum can_mode mode,
-				      u32_t bitrate)
-{
-	const struct can_driver_api *api = dev->driver_api;
-
-	return api->configure(dev, mode, bitrate);
-}
 
 #ifdef __cplusplus
 }
