+++ include/i2c.h
@@ -187,14 +187,6 @@ struct i2c_slave_driver_api {
  */
 __syscall int i2c_configure(struct device *dev, u32_t dev_config);
 
-static inline int _impl_i2c_configure(struct device *dev, u32_t dev_config)
-{
-	const struct i2c_driver_api *api =
-		(const struct i2c_driver_api *)dev->driver_api;
-
-	return api->configure(dev, dev_config);
-}
-
 /**
  * @brief Perform data transfer to another I2C device.
  *
@@ -216,16 +208,6 @@ static inline int _impl_i2c_configure(struct device *dev, u32_t dev_config)
 __syscall int i2c_transfer(struct device *dev,
 			   struct i2c_msg *msgs, u8_t num_msgs,
 			   u16_t addr);
-
-static inline int _impl_i2c_transfer(struct device *dev,
-				     struct i2c_msg *msgs, u8_t num_msgs,
-				     u16_t addr)
-{
-	const struct i2c_driver_api *api =
-		(const struct i2c_driver_api *)dev->driver_api;
-
-	return api->transfer(dev, msgs, num_msgs, addr);
-}
 
 /**
  * @brief Registers the provided config as Slave device
@@ -253,19 +235,6 @@ static inline int _impl_i2c_transfer(struct device *dev,
 __syscall int i2c_slave_register(struct device *dev,
 				 struct i2c_slave_config *cfg);
 
-static inline int _impl_i2c_slave_register(struct device *dev,
-					   struct i2c_slave_config *cfg)
-{
-	const struct i2c_driver_api *api =
-		(const struct i2c_driver_api *)dev->driver_api;
-
-	if (!api->slave_register) {
-		return -ENOTSUP;
-	}
-
-	return api->slave_register(dev, cfg);
-}
-
 /**
  * @brief Unregisters the provided config as Slave device
  *
@@ -284,19 +253,6 @@ static inline int _impl_i2c_slave_register(struct device *dev,
 __syscall int i2c_slave_unregister(struct device *dev,
 				   struct i2c_slave_config *cfg);
 
-static inline int _impl_i2c_slave_unregister(struct device *dev,
-					     struct i2c_slave_config *cfg)
-{
-	const struct i2c_driver_api *api =
-		(const struct i2c_driver_api *)dev->driver_api;
-
-	if (!api->slave_unregister) {
-		return -ENOTSUP;
-	}
-
-	return api->slave_unregister(dev, cfg);
-}
-
 /**
  * @brief Instructs the I2C Slave device to register itself to the I2C Controller
  *
@@ -311,14 +267,6 @@ static inline int _impl_i2c_slave_unregister(struct device *dev,
  */
 __syscall int i2c_slave_driver_register(struct device *dev);
 
-static inline int _impl_i2c_slave_driver_register(struct device *dev)
-{
-	const struct i2c_slave_driver_api *api =
-		(const struct i2c_slave_driver_api *)dev->driver_api;
-
-	return api->driver_register(dev);
-}
-
 /**
  * @brief Instructs the I2C Slave device to unregister itself from the I2C
  * Controller
@@ -332,14 +280,6 @@ static inline int _impl_i2c_slave_driver_register(struct device *dev)
  * @retval -EINVAL If parameters are invalid
  */
 __syscall int i2c_slave_driver_unregister(struct device *dev);
-
-static inline int _impl_i2c_slave_driver_unregister(struct device *dev)
-{
-	const struct i2c_slave_driver_api *api =
-		(const struct i2c_slave_driver_api *)dev->driver_api;
-
-	return api->driver_unregister(dev);
-}
 
 /*
  * Derived i2c APIs -- all implemented in terms of i2c_transfer()
