+++ include/ipm.h
@@ -131,14 +131,6 @@ struct ipm_driver_api {
 __syscall int ipm_send(struct device *ipmdev, int wait, u32_t id,
 		       const void *data, int size);
 
-static inline int _impl_ipm_send(struct device *ipmdev, int wait, u32_t id,
-			   const void *data, int size)
-{
-	const struct ipm_driver_api *api = ipmdev->driver_api;
-
-	return api->send(ipmdev, wait, id, data, size);
-}
-
 /**
  * @brief Register a callback function for incoming messages.
  *
@@ -167,14 +159,6 @@ static inline void ipm_register_callback(struct device *ipmdev,
  */
 __syscall int ipm_max_data_size_get(struct device *ipmdev);
 
-static inline int _impl_ipm_max_data_size_get(struct device *ipmdev)
-{
-	const struct ipm_driver_api *api = ipmdev->driver_api;
-
-	return api->max_data_size_get(ipmdev);
-}
-
-
 /**
  * @brief Return the maximum id value possible in an outbound message.
  *
@@ -187,13 +171,6 @@ static inline int _impl_ipm_max_data_size_get(struct device *ipmdev)
  */
 __syscall u32_t ipm_max_id_val_get(struct device *ipmdev);
 
-static inline u32_t _impl_ipm_max_id_val_get(struct device *ipmdev)
-{
-	const struct ipm_driver_api *api = ipmdev->driver_api;
-
-	return api->max_id_val_get(ipmdev);
-}
-
 /**
  * @brief Enable interrupts and callbacks for inbound channels.
  *
@@ -204,13 +181,6 @@ static inline u32_t _impl_ipm_max_id_val_get(struct device *ipmdev)
  * @retval EINVAL If it isn't an inbound channel.
  */
 __syscall int ipm_set_enabled(struct device *ipmdev, int enable);
-
-static inline int _impl_ipm_set_enabled(struct device *ipmdev, int enable)
-{
-	const struct ipm_driver_api *api = ipmdev->driver_api;
-
-	return api->set_enabled(ipmdev, enable);
-}
 
 #ifdef __cplusplus
 }
