+++ include/pwm.h
@@ -67,15 +67,6 @@ struct pwm_driver_api {
 __syscall int pwm_pin_set_cycles(struct device *dev, u32_t pwm,
 				 u32_t period, u32_t pulse);
 
-static inline int _impl_pwm_pin_set_cycles(struct device *dev, u32_t pwm,
-					   u32_t period, u32_t pulse)
-{
-	struct pwm_driver_api *api;
-
-	api = (struct pwm_driver_api *)dev->driver_api;
-	return api->pin_set(dev, pwm, period, pulse);
-}
-
 /**
  * @brief Get the clock rate (cycles per second) for a single PWM output.
  *
@@ -90,15 +81,6 @@ static inline int _impl_pwm_pin_set_cycles(struct device *dev, u32_t pwm,
 
 __syscall int pwm_get_cycles_per_sec(struct device *dev, u32_t pwm,
 				     u64_t *cycles);
-
-static inline int _impl_pwm_get_cycles_per_sec(struct device *dev, u32_t pwm,
-					       u64_t *cycles)
-{
-	struct pwm_driver_api *api;
-
-	api = (struct pwm_driver_api *)dev->driver_api;
-	return api->get_cycles_per_sec(dev, pwm, cycles);
-}
 
 /**
  * @brief Set the period and pulse width for a single PWM output.
