+++ include/rtc.h
@@ -59,30 +59,9 @@ struct rtc_driver_api {
 
 __syscall u32_t rtc_read(struct device *dev);
 
-static inline u32_t _impl_rtc_read(struct device *dev)
-{
-	const struct rtc_driver_api *api = dev->driver_api;
-
-	return api->read(dev);
-}
-
 __syscall void rtc_enable(struct device *dev);
 
-static inline void _impl_rtc_enable(struct device *dev)
-{
-	const struct rtc_driver_api *api = dev->driver_api;
-
-	api->enable(dev);
-}
-
 __syscall void rtc_disable(struct device *dev);
-
-static inline void _impl_rtc_disable(struct device *dev)
-{
-	const struct rtc_driver_api *api = dev->driver_api;
-
-	api->disable(dev);
-}
 
 static inline int rtc_set_config(struct device *dev,
 				 struct rtc_config *cfg)
@@ -93,14 +72,6 @@ static inline int rtc_set_config(struct device *dev,
 }
 
 __syscall int rtc_set_alarm(struct device *dev, const u32_t alarm_val);
-
-static inline int _impl_rtc_set_alarm(struct device *dev,
-				      const u32_t alarm_val)
-{
-	const struct rtc_driver_api *api = dev->driver_api;
-
-	return api->set_alarm(dev, alarm_val);
-}
 
 /**
  * @brief Function to get pending interrupts
@@ -116,14 +87,6 @@ static inline int _impl_rtc_set_alarm(struct device *dev,
  * @retval 0 if no rtc interrupt is pending.
  */
 __syscall int rtc_get_pending_int(struct device *dev);
-
-static inline int _impl_rtc_get_pending_int(struct device *dev)
-{
-	struct rtc_driver_api *api;
-
-	api = (struct rtc_driver_api *)dev->driver_api;
-	return api->get_pending_int(dev);
-}
 
 #ifdef __cplusplus
 }
