+++ include/spi.h
@@ -252,17 +252,6 @@ __syscall int spi_transceive(struct device *dev,
 			     const struct spi_buf_set *tx_bufs,
 			     const struct spi_buf_set *rx_bufs);
 
-static inline int _impl_spi_transceive(struct device *dev,
-				       const struct spi_config *config,
-				       const struct spi_buf_set *tx_bufs,
-				       const struct spi_buf_set *rx_bufs)
-{
-	const struct spi_driver_api *api =
-		(const struct spi_driver_api *)dev->driver_api;
-
-	return api->transceive(dev, config, tx_bufs, rx_bufs);
-}
-
 /**
  * @brief Read the specified amount of data from the SPI driver.
  *
@@ -402,15 +391,6 @@ static inline int spi_write_async(struct device *dev,
  */
 __syscall int spi_release(struct device *dev,
 			  const struct spi_config *config);
-
-static inline int _impl_spi_release(struct device *dev,
-				    const struct spi_config *config)
-{
-	const struct spi_driver_api *api =
-		(const struct spi_driver_api *)dev->driver_api;
-
-	return api->release(dev, config);
-}
 
 #ifdef __cplusplus
 }
